LLM Parameters Assignment - 02/requirements.txt
openai>=1.0.0
transformers>=4.30.0
jupyter>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.9.0
numpy>=1.21.0
pandas>=1.3.0